+++ app/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -3,20 +3,25 @@ package nextstep.jwp.controller;
 import java.util.NoSuchElementException;
 import java.util.UUID;
 
-import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.LoginException;
 import nextstep.jwp.handler.Cookie;
-import nextstep.jwp.handler.HttpBody;
 import nextstep.jwp.handler.HttpSession;
 import nextstep.jwp.handler.HttpSessions;
 import nextstep.jwp.handler.request.HttpRequest;
 import nextstep.jwp.handler.response.HttpResponse;
 import nextstep.jwp.model.User;
+import nextstep.jwp.service.LoginService;
 
 public class LoginController extends AbstractController {
 
 
     public static final String INDEX_HTML = "/index.html";
+
+    private final LoginService loginService;
+
+    public LoginController() {
+        this.loginService = new LoginService();
+    }
 
     @Override
     protected void doGet(HttpRequest httpRequest, HttpResponse httpResponse) {
@@ -42,15 +47,8 @@ public class LoginController extends AbstractController {
 
     @Override
     protected void doPost(HttpRequest httpRequest, HttpResponse httpResponse) {
-        HttpBody httpBody = httpRequest.getBody();
-        String account = httpBody.getBodyParams("account");
-        String password = httpBody.getBodyParams("password");
-
         try {
-            User user = findUser(account);
-            if (!user.checkPassword(password)) {
-                throw new LoginException("User의 정보와 입력한 정보가 일치하지 않습니다.");
-            }
+            User user = loginService.login(httpRequest);
 
             if (httpRequest.containsCookie("JSESSIONID")) {
                 final HttpSession session = httpRequest.getSession();
@@ -75,10 +73,5 @@ public class LoginController extends AbstractController {
         String sessionId = UUID.randomUUID().toString();
         HttpSessions.add(sessionId);
         return sessionId;
-    }
-
-    private User findUser(String account) {
-        return InMemoryUserRepository.findByAccount(account)
-                                     .orElseThrow(() -> { throw new LoginException("해당 User가 존재하지 않습니다."); });
     }
 }
